[probably not for upstream] CodeGen, libcxx: Introduce __has_non_relocatable_fields.
In order to properly benchmark PFP we need to be able to do an A/B
comparison and make A and B as equal as possible. This includes not
disabling libc++'s non-standard trivially relocatable optimizations
wherever possible. In the upstream code we just made trivially_relocatable
equivalent to trivially_copyable but in the benchmarking version of
the code (also more aligned to what will be possible in C++26) restore
trivially_relocatable logic but also check whether any of the fields
are non-relocatable using a temporarily introduced builtin.